+++ src/courseware/data/__factories__/courseBlocks.factory.js
@@ -35,3 +35,45 @@ Factory.define('courseBlocks')
     }),
   )
   .attr('root', ['course'], course => course.id);
+
+/**
+ * Builds a course with a single chapter, sequence, and unit.
+ */
+export default function buildSimpleCourseBlocks(courseId, title) {
+  const unitBlock = Factory.build(
+    'block',
+    { type: 'vertical' },
+    { courseId },
+  );
+  const sequenceBlock = Factory.build(
+    'block',
+    { type: 'sequential', children: [unitBlock.id] },
+    { courseId },
+  );
+  const sectionBlock = Factory.build(
+    'block',
+    { type: 'chapter', children: [sequenceBlock.id] },
+    { courseId },
+  );
+  const courseBlock = Factory.build(
+    'block',
+    { type: 'course', display_name: title, children: [sectionBlock.id] },
+    { courseId },
+  );
+  return {
+    courseBlocks: Factory.build(
+      'courseBlocks',
+      { courseId },
+      {
+        unit: unitBlock,
+        sequence: sequenceBlock,
+        section: sectionBlock,
+        course: courseBlock,
+      },
+    ),
+    unitBlock,
+    sequenceBlock,
+    sectionBlock,
+    courseBlock,
+  };
+}
